refactor(desktop): split demo build config
