+++ apps/desktop/vite.config.ts
@@ -2,49 +2,14 @@ import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import path from 'path';
 
-function withDemoTrailingSlash(rawUrl?: string | null): string | null {
-  if (!rawUrl) return null;
-  const [pathname, search = ''] = rawUrl.split('?', 2);
-  if (pathname === '/library-demo' || pathname === '/ui-demo') {
-    return `${pathname}/${search ? `?${search}` : ''}`;
-  }
-  return null;
-}
-
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [
-    react(),
-    {
-      name: 'demo-route-trailing-slash-redirect',
-      configureServer(server) {
-        server.middlewares.use((req, res, next) => {
-          const redirectTo = withDemoTrailingSlash(req.url);
-          if (!redirectTo) {
-            next();
-            return;
-          }
-
-          res.statusCode = 302;
-          res.setHeader('Location', redirectTo);
-          res.end();
-        });
-      },
-      configurePreviewServer(server) {
-        server.middlewares.use((req, res, next) => {
-          const redirectTo = withDemoTrailingSlash(req.url);
-          if (!redirectTo) {
-            next();
-            return;
-          }
-
-          res.statusCode = 302;
-          res.setHeader('Location', redirectTo);
-          res.end();
-        });
-      },
+  plugins: [react()],
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, 'src'),
     },
-  ],
+  },
   // Vite options tailored for Tauri development and only applied in `tauri dev` or `tauri build`
   // prevent vite from obscuring rust errors
   clearScreen: false,
@@ -66,8 +31,6 @@ export default defineConfig({
     rollupOptions: {
       input: {
         app: path.resolve(process.cwd(), 'index.html'),
-        libraryDemo: path.resolve(process.cwd(), 'library-demo/index.html'),
-        uiDemo: path.resolve(process.cwd(), 'ui-demo/index.html'),
       },
     },
   },
